+++ lib/error-handler.js
@@ -2,7 +2,6 @@
 
 const FJS = require('fast-json-stringify')
 const statusCodes = require('http').STATUS_CODES
-const setErrorHeaders = require('./setErrorHeaders')
 const wrapThenable = require('./wrapThenable')
 const {
   kReplyHeaders, kReplyNextErrorHandler, kReplyIsRunningOnErrorHook, kReplySent, kReplyHasStatusCode
@@ -28,12 +27,10 @@ const rootErrorHandler = {
 }
 
 function handleError (reply, error, cb) {
-  console.log('handleError')
   reply[kReplyIsRunningOnErrorHook] = false
 
   const context = reply.context
   if (reply[kReplyNextErrorHandler] === false) {
-    console.log('aaa')
     fallbackErrorHandler(error, reply, function (reply, payload) {
       reply.raw.writeHead(reply.raw.statusCode, reply[kReplyHeaders])
       reply.raw.end(payload)
@@ -42,13 +39,12 @@ function handleError (reply, error, cb) {
   }
   const errorHandler = reply[kReplyNextErrorHandler] || context.errorHandler
 
-  // In case we the error handler throws, we set the next errorHandler so we can error again
+  // In case the error handler throws, we set the next errorHandler so we can error again
   reply[kReplyNextErrorHandler] = Object.getPrototypeOf(errorHandler)
 
   reply[kReplyHeaders]['content-length'] = undefined
 
   const func = errorHandler.func
-  console.log(func)
 
   if (!func) {
     reply[kReplyNextErrorHandler] = false
@@ -56,7 +52,6 @@ function handleError (reply, error, cb) {
     return
   }
 
-  console.log('calling the error handler')
   const result = func(error, reply.request, reply)
   if (result !== undefined) {
     if (result !== null && typeof result.then === 'function') {
@@ -69,7 +64,6 @@ function handleError (reply, error, cb) {
 
 function defaultErrorHandler (error, request, reply) {
   setErrorHeaders(error, reply)
-  console.log('defaultErrorHandler', reply.statusCode, error.statusCode, reply[kReplyHasStatusCode])
   if (!reply[kReplyHasStatusCode] || reply.statusCode === 200) {
     const statusCode = error.statusCode || error.status
     reply.code(statusCode >= 400 ? statusCode : 500)
@@ -89,10 +83,8 @@ function defaultErrorHandler (error, request, reply) {
 }
 
 function fallbackErrorHandler (error, reply, cb) {
-  console.log('fallbackErrorHandler', reply.statusCode, error.statusCode, reply[kReplyHasStatusCode])
   const res = reply.raw
   const statusCode = reply.statusCode
-  console.log(statusCode)
   let payload
   try {
     const serializerFn = getSchemaSerializer(reply.context, statusCode)
@@ -124,15 +116,7 @@ function fallbackErrorHandler (error, reply, cb) {
 
   reply[kReplySent] = true
 
-  console.log('headerSent', res._headerSent)
-
-  console.log(cb)
   cb(reply, payload)
-
-  /*
-  res.writeHead(statusCode, reply[kReplyHeaders])
-  res.end(payload)
-  */
 }
 
 function buildErrorHandler (parent = rootErrorHandler, func) {
@@ -143,6 +127,24 @@ function buildErrorHandler (parent = rootErrorHandler, func) {
   const errorHandler = Object.create(parent)
   errorHandler.func = func
   return errorHandler
+}
+
+function setErrorHeaders (error, reply) {
+  const res = reply.raw
+  let statusCode = res.statusCode
+  statusCode = (statusCode >= 400) ? statusCode : 500
+  // treat undefined and null as same
+  if (error != null) {
+    if (error.headers !== undefined) {
+      reply.headers(error.headers)
+    }
+    if (error.status >= 400) {
+      statusCode = error.status
+    } else if (error.statusCode >= 400) {
+      statusCode = error.statusCode
+    }
+  }
+  res.statusCode = statusCode
 }
 
 module.exports = {
